update pl_pl.lang
The translation was made from the en_us.lang file to the BiblioCraft-Mod-1.12 version because the file that was in version 1.12.1 translated only 30% of the variables contained in the en_us.lang file and had incorrectly encoded diacritics that use the Polish language. .
I suspect this file should work correctly with all versions 1.12 but unfortunately I did not check the downgrade.